+++ spec/dummy/app/models/like.rb
@@ -1,0 +1,5 @@
+class Like < ActiveRecord::Base
+  attr_accessible :name, :user_id
+
+  belongs_to :user
+end