+++ spec/dummy/app/models/user.rb
@@ -2,4 +2,6 @@ require 'acts_as_xlsx'
 class User < ActiveRecord::Base
 	acts_as_xlsx columns: [:id, :name, :last_name, :address, :email]
   attr_accessible :address, :email, :last_name, :name
+
+  has_many :likes
 end