+++ spec/dummy/db/migrate/20121206210955_create_likes.rb
@@ -1,0 +1,10 @@
+class CreateLikes < ActiveRecord::Migration
+  def change
+    create_table :likes do |t|
+      t.string :name
+      t.integer :user_id
+
+      t.timestamps
+    end
+  end
+end